Revert "Remove debuglog option Google Play builds. (#171)"
This reverts commit 794ea0a0642ed0f4c4b48bc8aaab4bd2aace413c.
