+++ app/src/main/java/eu/darken/myperm/App.kt
@@ -6,14 +6,12 @@ import coil.ImageLoaderFactory
 import dagger.hilt.android.HiltAndroidApp
 import eu.darken.myperm.common.debug.autoreport.AutomaticBugReporter
 import eu.darken.myperm.common.debug.logging.*
-import eu.darken.myperm.common.debug.recording.core.RecorderModule
 import javax.inject.Inject
 
 @HiltAndroidApp
 open class App : Application() {
 
     @Inject lateinit var bugReporter: AutomaticBugReporter
-    @Inject lateinit var recorderModule: RecorderModule
     @Inject lateinit var imageLoaderFactory: ImageLoaderFactory
 
     override fun onCreate() {
