+++ app/src/main/java/eu/darken/myperm/settings/ui/support/SupportFragmentVM.kt
@@ -18,8 +18,6 @@ class SupportFragmentVM @Inject constructor(
 
     val isRecording = recorderModule.state.map { it.isRecording }.asLiveData2()
 
-    val isRecorderAvailable = recorderModule.state.map { it.isAvailable }.asLiveData2()
-
     fun startDebugLog() = launch {
         log { "startDebugLog()" }
         recorderModule.startRecorder()
