+++ app/src/main/res/xml/preferences_support.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <Preference
         android:icon="@drawable/ic_card_text_onsurface"
@@ -28,16 +27,10 @@
             android:data="https://discord.gg/7gGWxfM5yv" />
     </Preference>
 
-    <PreferenceCategory
-        android:key="category.other"
-        app:isPreferenceVisible="false"
-        tools:isPreferenceVisible="true"
-        app:title="@string/settings_category_other_label">
+    <PreferenceCategory app:title="@string/settings_category_other_label">
         <Preference
             android:icon="@drawable/ic_baseline_bug_report_24"
             android:key="support.debuglog"
-            app:isPreferenceVisible="false"
-            tools:isPreferenceVisible="true"
             android:summary="@string/support_debuglog_desc"
             android:title="@string/support_debuglog_label" />
     </PreferenceCategory>
